Move promotion code to lclmorph.cpp
