+++ src/coreclr/jit/lclmorph.cpp
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 #include "jitpch.h"
+#include "jitstd/algorithm.h"
 
 class LocalAddressVisitor final : public GenTreeVisitor<LocalAddressVisitor>
 {
@@ -2472,6 +2473,730 @@ private:
     }
 #endif
 };
+
+bool StructPromotionHelper::TryPromoteStructLocal(unsigned lclNum)
+{
+    if (CanPromoteStructLocal(lclNum) && ShouldPromoteStructLocal(lclNum))
+    {
+        PromoteStructLocal(lclNum);
+        return true;
+    }
+
+    return false;
+}
+
+bool StructPromotionHelper::CanPromoteStructType(CORINFO_CLASS_HANDLE typeHandle)
+{
+    // TODO-ObjectStackAllocation: Enable promotion of fields of stack-allocated objects.
+    assert(compiler->info.compCompHnd->isValueClass(typeHandle));
+
+    if (info.typeHandle == typeHandle)
+    {
+        // Asking for the same type of struct as the last time.
+        // Nothing need to be done.
+        return info.canPromoteStructType;
+    }
+
+    ICorJitInfo* vm = compiler->info.compCompHnd;
+
+    unsigned structSize = vm->getClassSize(typeHandle);
+
+    if (structSize > MaxStructSize)
+    {
+        return false;
+    }
+
+    unsigned fieldCount = vm->getClassNumInstanceFields(typeHandle);
+
+    if ((fieldCount < 1) || (fieldCount > MaxFieldCount))
+    {
+        return false;
+    }
+
+    uint32_t typeFlags = vm->getClassAttribs(typeHandle);
+
+    if ((typeFlags & CORINFO_FLG_DONT_PROMOTE) != 0)
+    {
+        // In AOT ReadyToRun compilation, don't try to promote fields of types
+        // outside of the current version bubble.
+        return false;
+    }
+
+    if ((typeFlags & CORINFO_FLG_OVERLAPPING_FIELDS) != 0)
+    {
+        return false;
+    }
+
+    if ((typeFlags & CORINFO_FLG_CUSTOMLAYOUT) != 0)
+    {
+        // Don't promote HFAs with custom layout.
+        // TODO-MIKE-Review: How exactly a HFA can have custom layout?!?
+
+        ClassLayout* layout = compiler->typGetObjLayout(typeHandle);
+        layout->EnsureHfaInfo(compiler);
+
+        if (layout->IsHfa())
+        {
+            return false;
+        }
+
+        // If we have "Custom Layout" then we might have an explicit Size attribute
+        // Managed C++ uses this for its structs, such C++ types will not contain GC pointers.
+        //
+        // The current VM implementation also incorrectly sets the CORINFO_FLG_CUSTOMLAYOUT
+        // whenever a managed value class contains any GC pointers.
+        // (See the comment for VMFLAG_NOT_TIGHTLY_PACKED in class.h)
+        //
+        // It is important to struct promote managed value classes that have GC pointers
+        // So we compute the correct value for "CustomLayout" here
+
+        if ((typeFlags & CORINFO_FLG_CONTAINS_GC_PTR) == 0)
+        {
+            info.customLayout = true;
+        }
+    }
+
+#ifdef TARGET_ARM
+    // On ARM, we have a requirement on the struct alignment; see below.
+    unsigned structAlignment = roundUp(vm->getClassAlignmentRequirement(typeHandle), TARGET_POINTER_SIZE);
+#endif
+
+    unsigned totalFieldSize     = 0;
+    bool     containsGCpointers = false;
+
+    new (&info) StructInfo(typeHandle, fieldCount);
+
+    for (unsigned index = 0; index < fieldCount; ++index)
+    {
+        GetFieldInfo(index);
+
+        const FieldInfo& field = info.fields[index];
+
+        if (field.type == TYP_STRUCT)
+        {
+            return false;
+        }
+
+        unsigned size = varTypeSize(field.type);
+
+        // The end offset for this field should never be larger than our structSize.
+        noway_assert(field.offset + size <= structSize);
+
+        if (size > 1)
+        {
+            unsigned alignment = varTypeSize(field.type);
+
+#ifdef FEATURE_SIMD
+            if (varTypeIsSIMD(field.type) && (field.layout->GetVectorKind() == VectorKind::Vector234))
+            {
+                // Vector2/3/4 doesn't have special alignment rules in the VM,
+                // it has only FLOAT fields so it's 4 byte aligned.
+                alignment = 4;
+            }
+#endif
+
+            if (field.offset % alignment != 0)
+            {
+                // The code in Compiler::genPushArgList that reconstitutes
+                // struct values on the stack from promoted fields expects
+                // those fields to be at their natural alignment.
+                return false;
+            }
+        }
+
+#ifdef TARGET_ARM
+        // On ARM, for struct types that don't use explicit layout, the alignment of the struct is
+        // at least the max alignment of its fields.  We take advantage of this invariant in struct
+        // promotion, so verify it here.
+        if (size > structAlignment)
+        {
+            return false;
+        }
+#endif
+
+        containsGCpointers |= varTypeIsGC(field.type);
+        totalFieldSize += size;
+    }
+
+    noway_assert(!containsGCpointers ||
+                 ((typeFlags & (CORINFO_FLG_CONTAINS_GC_PTR | CORINFO_FLG_CONTAINS_STACK_PTR)) != 0));
+
+    // If sizes do not match it means we have an overlapping fields or holes.
+    // Overlapping fields were rejected early, so here it can mean only holes.
+    info.containsHoles = totalFieldSize != structSize;
+
+    if (info.containsHoles && (fieldCount == 1))
+    {
+        // Single field structs can only have holes due to explicit size/layout.
+        // Don't bother promoting, they're rare and a potential source of bugs
+        // (especially considering that the most commun occurence of such structs
+        // are "fixed buffer" structs created by the C# compiler and those must
+        // not be promoted).
+        return false;
+    }
+
+    info.canPromoteStructType = true;
+    return true;
+}
+
+bool StructPromotionHelper::CanPromoteStructLocal(unsigned lclNum)
+{
+    LclVarDsc* lcl = compiler->lvaGetDesc(lclNum);
+
+    assert(varTypeIsStruct(lcl->GetType()));
+    assert(!lcl->IsPromoted());
+
+    // If this local is used by SIMD intrinsics, then we don't want to promote it.
+    // Note, however, that Vector2/3/4 local that are NOT used by SIMD intrinsics
+    // may be profitably promoted.
+    if (lcl->lvIsUsedInSIMDIntrinsic())
+    {
+        JITDUMP("  promotion of V%02u is disabled due to SIMD intrinsic uses\n", lclNum);
+        return false;
+    }
+
+#ifdef TARGET_ARM
+    if (lcl->IsParam())
+    {
+        // TODO-MIKE-CQ: Promote ARM struct params.
+        return false;
+    }
+#endif
+
+    // If GS stack reordering is enabled we may introduce shadow copies of parameters.
+    if (lcl->IsParam() && compiler->compGSReorderStackLayout)
+    {
+        JITDUMP("  promotion of param V%02u is disabled due to GS stack layout reordering\n", lclNum);
+        return false;
+    }
+
+    if (!compiler->lvaEnregMultiRegVars)
+    {
+        if (lcl->lvIsMultiRegArg)
+        {
+            JITDUMP("  promotion of V%02u is disabled because it is a multi-reg arg\n", lclNum);
+            return false;
+        }
+
+        if (lcl->lvIsMultiRegRet)
+        {
+            JITDUMP("  promotion of V%02u is disabled because it is a multi-reg return\n", lclNum);
+            return false;
+        }
+    }
+
+    // TODO-CQ: enable promotion for OSR locals
+    if (compiler->lvaIsOSRLocal(lclNum))
+    {
+        JITDUMP("  promotion of V%02u is disabled because it is an OSR local\n", lclNum);
+        return false;
+    }
+
+    CORINFO_CLASS_HANDLE typeHandle = lcl->GetLayout()->GetClassHandle();
+
+    if (!CanPromoteStructType(typeHandle))
+    {
+        return false;
+    }
+
+#ifdef WINDOWS_AMD64_ABI
+    // TODO-MIKE-CQ: Promoting single FP field structs almost works on x64, the main
+    // problem is the handling of method parameters and unfortunately the code that
+    // does that is a pile of garbage...
+    if (lcl->IsParam() && (info.fieldCount == 1) && varTypeIsFloating(info.fields[0].type))
+    {
+        JITDUMP("Not promoting param V%02u: struct has a single float/double field.\n", lclNum, info.fieldCount);
+        return false;
+    }
+#endif
+
+    if (!lcl->lvIsMultiRegArg && !lcl->lvIsMultiRegRet)
+    {
+        return true;
+    }
+
+    if (info.fieldCount > MAX_MULTIREG_COUNT)
+    {
+        return false;
+    }
+
+#if defined(TARGET_ARMARCH)
+    for (unsigned i = 0; i < info.fieldCount; i++)
+    {
+        const FieldInfo& field = info.fields[i];
+
+        // Non-HFA structs are always passed in general purpose registers.
+        // If there are any floating point fields, don't promote for now.
+        // TODO-1stClassStructs: add support in Lowering and prolog generation
+        // to enable promoting these types.
+        if (lcl->IsParam() && !lcl->lvIsHfa() && varTypeUsesFloatReg(field.type))
+        {
+            return false;
+        }
+
+#ifdef FEATURE_SIMD
+        // If we have a register-passed struct with mixed non-opaque SIMD types (i.e. with defined fields)
+        // and non-SIMD types, we don't currently handle that case in the prolog, so we can't promote.
+        if ((info.fieldCount > 1) && varTypeIsStruct(field.type) && !field.layout->IsOpaqueVector())
+        {
+            return false;
+        }
+#endif
+    }
+#elif defined(UNIX_AMD64_ABI)
+    // Only promote if the field types match the registers, unless we have a single SIMD field
+    // that we can handle in prolog.
+
+    if ((info.fieldCount == 1) && varTypeIsSIMD(info.fields[0].type))
+    {
+        return true;
+    }
+
+    ClassLayout* layout = lcl->GetLayout();
+    layout->EnsureSysVAmd64AbiInfo(compiler);
+
+    if (info.fieldCount != layout->GetSysVAmd64AbiRegCount())
+    {
+        return false;
+    }
+
+    SortFields();
+
+    for (unsigned i = 0; i < info.fieldCount; i++)
+    {
+        const FieldInfo& field = info.fields[i];
+
+        // We don't currently support passing SIMD types in registers.
+        if (varTypeIsSIMD(field.type))
+        {
+            return false;
+        }
+
+        if (varTypeUsesFloatReg(field.type) != varTypeUsesFloatReg(layout->GetSysVAmd64AbiRegType(i)))
+        {
+            return false;
+        }
+    }
+#endif // UNIX_AMD64_ABI
+
+    return true;
+}
+
+bool StructPromotionHelper::ShouldPromoteStructLocal(unsigned lclNum)
+{
+    LclVarDsc* lcl = compiler->lvaGetDesc(lclNum);
+
+    assert(varTypeIsStruct(lcl->GetType()));
+    assert(lcl->GetLayout()->GetClassHandle() == info.typeHandle);
+    assert(info.canPromoteStructType);
+
+    // We *can* promote; *should* we promote?
+    //
+    // We should only do so if promotion has potential savings. One source of savings
+    // is if a field of the struct is accessed, since this access will be turned into
+    // an access of the corresponding promoted field variable. Even if there are no
+    // field accesses, but only block-level operations on the whole struct, if the struct
+    // has only one or two fields, then doing those block operations field-wise is probably
+    // faster than doing a whole-variable block operation (e.g., REP MOVSB).
+    //
+    // Struct promotion also provides the following benefits: reduce stack frame size,
+    // reduce the need for zero init of stack frame and fine grained constant/copy prop.
+    // Asm diffs indicate that promoting structs up to 3 fields is a net size win.
+    // So if no fields are accessed independently, and there are four or more fields,
+    // then do not promote.
+
+    // TODO: Ideally we would want to consider the impact of whether the struct is
+    // passed as a parameter or assigned the return value of a call. Because once promoted,
+    // struct copying is done by field by field assignment instead of a more efficient
+    // rep.stos or xmm reg based copy.
+
+    // TODO: If the lvRefCnt is zero and we have a struct promoted parameter we can end up
+    // with an extra store of the the incoming register into the stack frame slot.
+    // In that case, we would like to avoid promortion.
+    // However we haven't yet computed the lvRefCnt values so we can't do that.
+
+    // TODO-MIKE-CQ: SIMD promotion is still messy as lvFieldAccessed is way too limited.
+    // A few "get"/"set"s may be better expressed as "get element"/"with element". Repeated
+    // "get"s may be best handled by CSE. Only repeated "set"s are likely to benefit from
+    // promotion, especially when we don't have SSE41.
+    //
+    // On ARM64 it may be useful to treat Vector2/3/4 params and locals that are used by
+    // a RETURN as "field accessed" (if they don't have intrinsic uses). Otherwise we end
+    // up with pretty stupid codegen in some rather trivial cases (e.g. Vector3 param that
+    // is immediately passed as an arg to another call - currently we pack all the param regs
+    // into a single one in the prolog and then unpack the single reg to multiple regs before
+    // the call). But if the param is passed as an argument to multiple calls then things
+    // are more complicated, it may better to not promote to avoid unnecessary register
+    // pressure. The current struct promotion is way too limited to be able to do the right
+    // thing in this case.
+
+    if (!lcl->lvFieldAccessed && (info.fieldCount > 3 || varTypeIsSIMD(lcl->GetType())))
+    {
+        JITDUMP("Not promoting V%02u: type = %s, #fields = %d, fieldAccessed = %d.\n", lclNum,
+                varTypeName(lcl->GetType()), info.fieldCount, lcl->lvFieldAccessed);
+        return false;
+    }
+
+    if (lcl->lvIsMultiRegRet && info.containsHoles && info.customLayout)
+    {
+        JITDUMP("Not promoting multi-reg returned V%02u with holes.\n", lclNum);
+        return false;
+    }
+
+    if (lcl->IsParam() && !lcl->IsImplicitByRefParam() && !lcl->lvIsHfa())
+    {
+#if FEATURE_MULTIREG_STRUCT_PROMOTE
+        if (compiler->lvaIsMultiRegStructParam(lcl))
+        {
+            if ((info.fieldCount != 2) && ((info.fieldCount != 1) || !varTypeIsSIMD(info.fields[0].type)))
+            {
+                JITDUMP("Not promoting multireg param V%02u, #fields != 2 and it's not SIMD.\n", lclNum);
+                return false;
+            }
+        }
+        else
+#endif // !FEATURE_MULTIREG_STRUCT_PROMOTE
+
+            // TODO-PERF - Implement struct promotion for incoming single-register structs.
+            //             Also the implementation of jmp uses the 4 byte move to store
+            //             byte parameters to the stack, so that if we have a byte field
+            //             with something else occupying the same 4-byte slot, it will
+            //             overwrite other fields.
+            if (info.fieldCount != 1)
+        {
+            JITDUMP("Not promoting param V%02u, #fields = %u.\n", lclNum, info.fieldCount);
+            return false;
+        }
+    }
+
+    if ((lclNum == compiler->genReturnLocal) && (info.fieldCount > 1))
+    {
+        // TODO-1stClassStructs: a temporary solution to keep diffs small, it will be fixed later.
+        return false;
+    }
+
+#ifdef DEBUG
+    if (compiler->compPromoteFewerStructs(lclNum))
+    {
+        JITDUMP("Not promoting promotable V%02u, because of STRESS_PROMOTE_FEWER_STRUCTS\n", lclNum);
+        return false;
+    }
+#endif
+
+    return true;
+}
+
+void StructPromotionHelper::SortFields()
+{
+    if (!info.fieldsSorted)
+    {
+        info.fieldsSorted = true;
+
+        jitstd::sort(info.fields, info.fields + info.fieldCount,
+                     [](const FieldInfo& f1, const FieldInfo& f2) { return f1.offset < f2.offset; });
+    }
+}
+
+void StructPromotionHelper::GetFieldInfo(unsigned index)
+{
+    ICorJitInfo* vm = compiler->info.compCompHnd;
+
+    CORINFO_FIELD_HANDLE fieldHandle = vm->getFieldInClass(info.typeHandle, index);
+    CORINFO_CLASS_HANDLE typeHandle;
+    var_types            type = JITtype2varType(vm->getFieldType(fieldHandle, &typeHandle));
+
+    FieldInfo& field     = info.fields[index];
+    field.fieldSeq[0]    = fieldHandle;
+    field.fieldSeqLength = 1;
+    field.offset         = vm->getFieldOffset(fieldHandle);
+    field.type           = type;
+    field.layout         = nullptr;
+
+    if (field.type == TYP_STRUCT)
+    {
+#ifdef FEATURE_SIMD
+        ClassLayout* layout = compiler->supportSIMDTypes() ? compiler->typGetObjLayout(typeHandle) : nullptr;
+
+        if ((layout != nullptr) && layout->IsVector())
+        {
+            field.type   = layout->GetSIMDType();
+            field.layout = layout;
+        }
+        else
+#endif
+        {
+            GetSingleFieldStructInfo(field, typeHandle);
+        }
+    }
+}
+
+void StructPromotionHelper::GetSingleFieldStructInfo(FieldInfo& field, CORINFO_CLASS_HANDLE typeHandle)
+{
+    assert(field.type == TYP_STRUCT);
+    assert(field.fieldSeqLength == 1);
+    assert(typeHandle != NO_CLASS_HANDLE);
+
+    ICorJitInfo* vm = compiler->info.compCompHnd;
+
+    uint8_t      depth  = 1;
+    var_types    type   = TYP_STRUCT;
+    ClassLayout* layout = nullptr;
+    unsigned     size   = 0;
+
+    while ((depth < _countof(field.fieldSeq)) && (type == TYP_STRUCT))
+    {
+        // TODO-MIKE-Review: Should we check for CORINFO_FLG_DONT_PROMOTE?
+
+        if (vm->getClassNumInstanceFields(typeHandle) != 1)
+        {
+            return;
+        }
+
+        CORINFO_FIELD_HANDLE fieldHandle = vm->getFieldInClass(typeHandle, 0);
+        CORINFO_CLASS_HANDLE fieldTypeHandle;
+        type   = CorTypeToVarType(vm->getFieldType(fieldHandle, &fieldTypeHandle));
+        size   = varTypeSize(type);
+        layout = nullptr;
+
+        if (type == TYP_STRUCT)
+        {
+            layout = compiler->typGetObjLayout(fieldTypeHandle);
+            size   = layout->GetSize();
+
+            if (layout->IsVector())
+            {
+                type = layout->GetSIMDType();
+            }
+        }
+
+        if ((vm->getFieldOffset(fieldHandle) != 0) || (vm->getClassSize(typeHandle) != size))
+        {
+            JITDUMP("Promotion blocked: single field struct contains padding\n");
+            return;
+        }
+
+        field.fieldSeq[depth++] = fieldHandle;
+        typeHandle              = fieldTypeHandle;
+    }
+
+    if (type != TYP_STRUCT)
+    {
+        field.fieldSeqLength = depth;
+        field.type           = type;
+        field.layout         = layout;
+    }
+}
+
+void StructPromotionHelper::PromoteStructLocal(unsigned lclNum)
+{
+    LclVarDsc* lcl = compiler->lvaGetDesc(lclNum);
+
+    // We should never see a reg-sized non-field-addressed struct here.
+    assert(!lcl->lvRegStruct);
+
+    assert(lcl->GetLayout()->GetClassHandle() == info.typeHandle);
+    assert(info.canPromoteStructType);
+
+    lcl->lvFieldCnt      = info.fieldCount;
+    lcl->lvFieldLclStart = compiler->lvaCount;
+    lcl->lvPromoted      = true;
+    lcl->lvContainsHoles = info.containsHoles;
+    lcl->lvCustomLayout  = info.customLayout;
+
+    INDEBUG(lcl->lvKeepType = 1;)
+
+    JITDUMP("\nPromoting struct local V%02u (%s):", lclNum, lcl->GetLayout()->GetClassName());
+
+    SortFields();
+
+    for (unsigned index = 0; index < info.fieldCount; ++index)
+    {
+        const FieldInfo& field = info.fields[index];
+        assert((index == 0) || (field.offset > info.fields[index - 1].offset));
+
+        if (varTypeUsesFloatReg(field.type))
+        {
+            compiler->compFloatingPointUsed = true;
+        }
+        else if (field.type == TYP_LONG)
+        {
+            compiler->compLongUsed = true;
+        }
+
+        FieldSeqNode* fieldSeq = nullptr;
+
+        for (unsigned i = 0; i < field.fieldSeqLength; i++)
+        {
+            FieldSeqNode* fieldSeqNode = compiler->GetFieldSeqStore()->CreateSingleton(field.fieldSeq[i]);
+            fieldSeq                   = compiler->GetFieldSeqStore()->Append(fieldSeq, fieldSeqNode);
+        }
+
+        unsigned fieldLclNum = compiler->lvaGrabTemp(false DEBUGARG("promoted struct field"));
+        // lvaGrabTemp can reallocate the lvaTable, so refresh the cached lcl for lclNum.
+        lcl = compiler->lvaGetDesc(lclNum);
+
+        LclVarDsc* fieldLcl = compiler->lvaGetDesc(fieldLclNum);
+        fieldLcl->MakePromotedStructField(lclNum, field.offset, fieldSeq);
+
+        if (varTypeIsSIMD(field.type))
+        {
+            compiler->lvaSetStruct(fieldLclNum, field.layout, false);
+            // We will not recursively promote this, so mark it as 'lvRegStruct' (note that we wouldn't
+            // be promoting this if we didn't think it could be enregistered.
+            fieldLcl->lvRegStruct = true;
+        }
+        else
+        {
+            fieldLcl->SetType(field.type);
+        }
+
+        fieldLcl->lvIsParam = lcl->IsParam();
+
+        if (!lcl->IsRegParam())
+        {
+            continue;
+        }
+
+        fieldLcl->lvIsRegArg = true;
+
+#if !FEATURE_MULTIREG_ARGS
+        fieldLcl->SetArgReg(lcl->GetArgReg());
+#else
+        if (lcl->IsImplicitByRefParam())
+        {
+            fieldLcl->SetArgReg(lcl->GetArgReg());
+            continue;
+        }
+
+#ifdef UNIX_AMD64_ABI
+        if (varTypeIsSIMD(fieldLcl->GetType()) && (lcl->GetPromotedFieldCount() == 1))
+        {
+            fieldLcl->SetArgReg(lcl->GetArgReg());
+            fieldLcl->SetOtherArgReg(lcl->GetOtherArgReg());
+            continue;
+        }
+#endif
+
+        if (lcl->lvIsHfa())
+        {
+            // We've sorted the field by offset so for HFA's we should have a 1:1 mapping
+            // between fields and registers. Also, ensure that we didn't promote a HFA
+            // with holes. Is that even a thing? The VM seems to think that it is...
+            assert(field.offset == index * varTypeSize(lcl->GetLayout()->GetHfaElementType()));
+
+            unsigned regIndex = index;
+
+#ifdef TARGET_ARM
+            if (lcl->GetLayout()->GetHfaElementType() == TYP_DOUBLE)
+            {
+                // On ARM we count FLOAT rather than DOUBLE registers.
+                regIndex *= 2;
+            }
+#endif
+
+            fieldLcl->SetArgReg(static_cast<regNumber>(lcl->GetArgReg() + regIndex));
+            continue;
+        }
+
+        // TODO-ARMARCH: Need to determine if/how to handle split args.
+
+        if (index == 0)
+        {
+            fieldLcl->SetArgReg(lcl->GetArgReg());
+        }
+        else
+        {
+            assert(index == 1);
+            fieldLcl->SetArgReg(lcl->GetOtherArgReg());
+        }
+#endif // FEATURE_MULTIREG_ARGS
+    }
+}
+
+void Compiler::fgPromoteStructs()
+{
+    JITDUMP("*************** In fgPromoteStructs()\n");
+
+    if (!opts.OptEnabled(CLFLG_STRUCTPROMOTE))
+    {
+        JITDUMP("  promotion opt flag not enabled\n");
+        return;
+    }
+
+    if (fgNoStructPromotion)
+    {
+        JITDUMP("  promotion disabled by JitNoStructPromotion\n");
+        return;
+    }
+
+    if (info.compIsVarArgs)
+    {
+        JITDUMP("  promotion disabled because of varargs\n");
+        return;
+    }
+
+#ifdef DEBUG
+    if (verbose)
+    {
+        printf("\nlvaTable before fgPromoteStructs\n");
+        lvaTableDump();
+    }
+#endif
+
+    StructPromotionHelper helper(this);
+
+    for (unsigned lclNum = 0, lclCount = lvaCount; lclNum < lclCount; lclNum++)
+    {
+        if (lvaHaveManyLocals())
+        {
+            JITDUMP("Stopped promoting struct fields, due to too many locals.\n");
+            break;
+        }
+
+        LclVarDsc* lcl = lvaGetDesc(lclNum);
+
+        if (!varTypeIsStruct(lcl->GetType()))
+        {
+            continue;
+        }
+
+#ifdef DEBUG
+        if (lcl->IsParam() && fgNoStructParamPromotion)
+        {
+            continue;
+        }
+#endif
+
+        if (varTypeIsSIMD(lcl->GetType()) && (lcl->lvIsUsedInSIMDIntrinsic() || lcl->GetLayout()->IsOpaqueVector()))
+        {
+            // If we have marked this as lvUsedInSIMDIntrinsic, then we do not want to promote
+            // its fields. Instead, we will attempt to enregister the entire struct.
+            lcl->lvRegStruct = true;
+            continue;
+        }
+
+        if (!helper.TryPromoteStructLocal(lclNum))
+        {
+            // If we don't promote then lvIsMultiRegRet is meaningless.
+            lcl->lvIsMultiRegRet = false;
+
+            if (varTypeIsSIMD(lcl->GetType()) && !lcl->lvFieldAccessed)
+            {
+                // Even if we have not used this in a SIMD intrinsic, if it is not being promoted,
+                // we will treat it as a reg struct.
+                lcl->lvRegStruct = true;
+            }
+        }
+    }
+
+#ifdef DEBUG
+    if (verbose)
+    {
+        printf("\nlvaTable after fgPromoteStructs\n");
+        lvaTableDump();
+    }
+#endif
+}
 
 //------------------------------------------------------------------------
 // fgMarkAddressExposedLocals: Traverses the entire method and marks address
