+++ src/coreclr/jit/morph.cpp
@@ -16074,91 +16074,6 @@ void Compiler::fgPostExpandQmarkChecks()
 }
 #endif
 
-void Compiler::fgPromoteStructs()
-{
-    JITDUMP("*************** In fgPromoteStructs()\n");
-
-    if (!opts.OptEnabled(CLFLG_STRUCTPROMOTE))
-    {
-        JITDUMP("  promotion opt flag not enabled\n");
-        return;
-    }
-
-    if (fgNoStructPromotion)
-    {
-        JITDUMP("  promotion disabled by JitNoStructPromotion\n");
-        return;
-    }
-
-    if (info.compIsVarArgs)
-    {
-        JITDUMP("  promotion disabled because of varargs\n");
-        return;
-    }
-
-#ifdef DEBUG
-    if (verbose)
-    {
-        printf("\nlvaTable before fgPromoteStructs\n");
-        lvaTableDump();
-    }
-#endif
-
-    StructPromotionHelper helper(this);
-
-    for (unsigned lclNum = 0, lclCount = lvaCount; lclNum < lclCount; lclNum++)
-    {
-        if (lvaHaveManyLocals())
-        {
-            JITDUMP("Stopped promoting struct fields, due to too many locals.\n");
-            break;
-        }
-
-        LclVarDsc* lcl = lvaGetDesc(lclNum);
-
-        if (!varTypeIsStruct(lcl->GetType()))
-        {
-            continue;
-        }
-
-#ifdef DEBUG
-        if (lcl->IsParam() && fgNoStructParamPromotion)
-        {
-            continue;
-        }
-#endif
-
-        if (varTypeIsSIMD(lcl->GetType()) && (lcl->lvIsUsedInSIMDIntrinsic() || lcl->GetLayout()->IsOpaqueVector()))
-        {
-            // If we have marked this as lvUsedInSIMDIntrinsic, then we do not want to promote
-            // its fields. Instead, we will attempt to enregister the entire struct.
-            lcl->lvRegStruct = true;
-            continue;
-        }
-
-        if (!helper.TryPromoteStructLocal(lclNum))
-        {
-            // If we don't promote then lvIsMultiRegRet is meaningless.
-            lcl->lvIsMultiRegRet = false;
-
-            if (varTypeIsSIMD(lcl->GetType()) && !lcl->lvFieldAccessed)
-            {
-                // Even if we have not used this in a SIMD intrinsic, if it is not being promoted,
-                // we will treat it as a reg struct.
-                lcl->lvRegStruct = true;
-            }
-        }
-    }
-
-#ifdef DEBUG
-    if (verbose)
-    {
-        printf("\nlvaTable after fgPromoteStructs\n");
-        lvaTableDump();
-    }
-#endif
-}
-
 //------------------------------------------------------------------------
 // fgAddFieldSeqForZeroOffset:
 //    Associate a fieldSeq (with a zero offset) with the GenTree node 'addr'
